initial MVP of OpenAPI support
Very basic support for adding some openapi spec in config, and having it
exposed through the gateway.

Currentlt supported: list_tools and call_tools, with a very, very small
portion of the specification translated. Remaining work will be isolated
changes into the translation spec; hoping to merge this now to get the
parts that touch the rest of the codebase merged in